Fix cache streaming refactor follow-ups
